Move apiClient related stuff to core module and refactor user settings api
