--- src/types.ts
@@ -1,6 +1,0 @@
-export type ApiResponse<T> = {
-  data: T;
-  success: boolean;
-  message?: string;
-  invalidFields?: { [key: string]: string[] };
-};
